Maps: Add generic tool system
The Maps application now has a tool system. On the right, we now display
a tool strip, where the user can select one tool. This tool then
receives all UI events from the main MapWidget and can do things with
the event data. It also gets called at the end of the paint_event, to be
able to draw something on top of the map.

Tools are registered, similarly to PixelPaint's tool system, in a widget
called ToolStripWidget. This widget manages the tools and which tool is
active. In main.cpp, we add the tool to the main HorizontalSplitter.